core/requirements_updated.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI Clients
openai==1.10.0
httpx==0.26.0
tenacity==8.2.3

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Cache
redis[hiredis]==5.0.1

# Storage - Support both S3 and Supabase
aioboto3==12.3.0
supabase==2.3.0
minio==7.2.3

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6

# Image Processing (for asset generation)
Pillow==10.2.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.40.0

# WebSocket
python-socketio==5.11.0
websockets==12.0

# Utils
python-dotenv==1.0.0
python-dateutil==2.8.2
